refactor: reorgainze components folder
