+++ packages/takudev-web/components/atoms/index.ts
@@ -1,6 +1,6 @@
-export { CloseButton } from './CloseButton';
-export { HamburgerButton } from './HamburgerButton';
-export { Logo } from './Logo';
-export { ProfileIcon } from './ProfileIcon';
-export { SnsIcon } from './SnsIcon';
-export { TextTag } from './TextTag';
+export * from './CloseButton';
+export * from './HamburgerButton';
+export * from './Logo';
+export * from './ProfileIcon';
+export * from './SnsIcon';
+export * from './TextTag';
