+++ packages/takudev-web/components/features/Home/HomePageContent.module.scss
