+++ packages/takudev-web/components/features/Home/index.ts
@@ -1,0 +1,1 @@
+export { HomePageContent } from './HomePageContent';
