+++ packages/takudev-web/components/features/index.ts
@@ -1,0 +1,1 @@
+export * from './Home';
