+++ packages/takudev-web/components/organisms/index.ts
@@ -1,4 +1,4 @@
-export { PageHeader } from './PageHeader';
-export { PageFooter } from './PageFooter';
-export { PageContentHome } from './PageContentHome';
-export { ProfileCard } from './ProfileCard';
+export * from './ArticleSummaryCard';
+export * from './PageHeader';
+export * from './PageFooter';
+export * from './ProfileCard';
