--- packages/takudev-web/components/pages/Home/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { PageContentHome } from '~/components/organisms';
-import { PageTemplate } from '~/components/templates';
-
-export const HomePage: React.FC = () => (
-  <PageTemplate>
-    <PageContentHome />
-  </PageTemplate>
-);
