--- packages/takudev-web/components/pages/index.ts
@@ -1,1 +1,0 @@
-export { HomePage } from './Home';
