+++ packages/takudev-web/components/templates/index.ts
@@ -1,1 +1,1 @@
-export { PageTemplate } from './PageTemplate';
+export * from './PageTemplate';
