+++ packages/takudev-web/pages/index.tsx
@@ -1,16 +1,19 @@
 import Head from 'next/head';
-import { HomePage } from '~/components';
+import { HomePageContent } from '~/components/features';
+import { PageTemplate } from '~/components/templates';
 import type { NextPage } from 'next';
 
-const Home: NextPage = () => (
+const HomePage: NextPage = () => (
   <>
     <Head>
       <title>Taku.dev</title>
       <meta name='description' content="This is Takuya Iwashiro's Dev Blog" />
     </Head>
 
-    <HomePage />
+    <PageTemplate>
+      <HomePageContent />
+    </PageTemplate>
   </>
 );
 
-export default Home;
+export default HomePage;
